test(controlplane): split hosted genesis bootstrap cases
Move the hosted instance-trust genesis finalize/read regression tests into a focused package-local test file so MAI-1 file-budget enforcement stays green without changing coverage or production behavior.
